requirements.txt
pocketflow>=0.0.1
pandas>=2.0.0
PyYAML>=6.0
openai>=1.0.0
google-genai